+++ src/test/java/testclasses/casts/CastInt2Double.java
@@ -9,7 +9,7 @@ public class CastInt2Double {
 
     @NativeObfuscation
     public double exec(int a) {
-        return (double) (a + a);
+        return a + a;
     }
 
 }
