+++ src/test/java/testclasses/exceptions/invoke/invokestatic/CatchInvokeStaticLong.java
@@ -9,7 +9,7 @@ public class CatchInvokeStaticLong {
     }
 
     public static long normalDiv(int a, int b) {
-        return (long) (a / b);
+        return a / b;
     }
 
     @NativeObfuscation
