+++ src/test/java/testclasses/exceptions/invoke/invokevirtual/CatchInvokeVirtualDouble.java
@@ -27,7 +27,7 @@ public class CatchInvokeVirtualDouble {
     }
 
     public double normalDiv(int a, int b) {
-        return (double) (a / b);
+        return a / b;
     }
 
     public void other() throws UserDefinedException {
