+++ src/test/java/testclasses/exceptions/invoke/invokevirtual/ThrowInvokeVirtualLong.java
@@ -13,6 +13,6 @@ public class ThrowInvokeVirtualLong {
     }
 
     public long normalDiv(int a, int b) {
-        return (long) (a / b);
+        return a / b;
     }
 }
