+++ src/test/java/testclasses/exceptions/nullpointer/arrayload/CatchNullPointerExceptionXLOADByte.java
@@ -14,28 +14,28 @@ public class CatchNullPointerExceptionXLOADByte {
 
         int res = 0;
         try {
-            res += (int) obj.b[1];
+            res += obj.b[1];
             res += 1000;
         } catch (NullPointerException e) {
             res++;
         }
 
         try {
-            res += (int) obj.b[1];
+            res += obj.b[1];
             res += 1000;
         } catch (RuntimeException e) {
             res++;
         }
 
         try {
-            res += (int) obj.b[1];
+            res += obj.b[1];
             res += 1000;
         } catch (Exception e) {
             res++;
         }
 
         try {
-            res += (int) obj.b[1];
+            res += obj.b[1];
             res += 1000;
         } catch (Throwable e) {
             res++;
