+++ src/test/java/testclasses/exceptions/outofbounds/arrayload/CatchArrayLoadOutOfBoundsByte.java
@@ -14,35 +14,35 @@ public class CatchArrayLoadOutOfBoundsByte {
         int res = 0;
 
         try {
-            res += (int) this.array[2];
+            res += this.array[2];
             res += 1000;
         } catch (ArrayIndexOutOfBoundsException e) {
             res++;
         }
 
         try {
-            res += (int) this.array[2];
+            res += this.array[2];
             res += 1000;
         } catch (IndexOutOfBoundsException e) {
             res++;
         }
 
         try {
-            res += (int) this.array[2];
+            res += this.array[2];
             res += 1000;
         } catch (RuntimeException e) {
             res++;
         }
 
         try {
-            res += (int) this.array[2];
+            res += this.array[2];
             res += 1000;
         } catch (Exception e) {
             res++;
         }
 
         try {
-            res += (int) this.array[2];
+            res += this.array[2];
             res += 1000;
         } catch (Throwable e) {
             res++;
