+++ src/test/java/testclasses/exceptions/outofbounds/arraystore/CatchArrayStoreOutOfBoundsDouble.java
@@ -48,6 +48,6 @@ public class CatchArrayStoreOutOfBoundsDouble {
             res++;
         }
 
-        return (double) res;
+        return res;
     }
 }
