+++ src/test/java/testclasses/exceptions/outofbounds/arraystore/CatchArrayStoreOutOfBoundsLong.java
@@ -48,6 +48,6 @@ public class CatchArrayStoreOutOfBoundsLong {
             res++;
         }
 
-        return (long) res;
+        return res;
     }
 }
